Unnest impersonations controller now
Since it doesn't apply only to managed users.
