+++ decidim-admin/config/locales/en.yml
@@ -278,32 +278,32 @@ en:
       logs:
         logs_list:
           no_logs_yet: There are no logs yet
-      managed_users:
+      impersonations:
         close_session:
           error: There has been an error closing the current impersonation session.
           success: The current impersonation session has been successfully ended.
-        impersonations:
-          create:
-            error: There has been an error impersonating the user.
-            success: The managed user has been successfully created.
-          form:
-            authorization_method: Authorization method
-            name: Name
-            reason: Reason
-          index:
-            filter:
-              all: All
-              managed: Managed
-              not_managed: Not managed
-            filter_by: Filter by
-            impersonate: Impersonate
-            needs_authorization_warning: You need at least one authorization enabled for this organization.
-            promote: Promote
-            view_logs: View logs
-          new:
-            impersonate: Impersonate
-            new_managed_user: New managed user
-            new_managed_user_impersonation: New managed user impersonation
+        create:
+          error: There has been an error impersonating the user.
+          success: The managed user has been successfully created.
+        form:
+          authorization_method: Authorization method
+          name: Name
+          reason: Reason
+        index:
+          filter:
+            all: All
+            managed: Managed
+            not_managed: Not managed
+          filter_by: Filter by
+          impersonate: Impersonate
+          needs_authorization_warning: You need at least one authorization enabled for this organization.
+          promote: Promote
+          view_logs: View logs
+        new:
+          impersonate: Impersonate
+          new_managed_user: New managed user
+          new_managed_user_impersonation: New managed user impersonation
+      managed_users:
         promotion:
           error: There has been an error promoting the managed user.
           success: The managed user has been successfully promoted.
